Fixes #75 - Remove use of Client.cuo
- Client.cuo is deprecated and not used anymore. Use OrionUO.cfg.

Client.cuo is more of trouble than useful. This will make all needed
configuration via OrionUO.cfg file and soon configurable via the
launcher itself instead yet another closed source component.

This change also refactor how the client version identifier is built up
by using a uint32_t composition for each element of the client version
in one of the two possibilities:
1. major.minor.revision.prototype
2. major.minor.revision(letter)

The subrevision letter gets converted in a prototype field with the
decimal value of the letter itself.

This should also fix #63 as it mostly refactor how the crypto keys are
calculated and used and by eliminating the crypto Install that converted
the byte buffer into another so orion could use, which wasn't ideal
anyway.

This changes the flow of loading and saving OrionUO.cfg as it is now
required to be one of the first thigs to run, even before command line
parsing that will be able to override configuration values in the
future.