#57 Local file processing should depend on local configuration not on configuration sent by server.
It seems that it is because Sphere 0.99 behavior, it sends wrong value in `EnableLockedFeatures` when user is still in game when attempting to log in. So `g_LockedClientFeatures & LFF_LBR` is zero in `CAnimationManager::UpdateAnimationAddressTable`. Which means that `PatchedAddress` is not assigned to `Address` and `ReadFramesPixelData` access wrong memory when it loads animation data.